requirements.txt
flask==3.0.3
flask-cors==4.0.1
opencv-python==4.10.0.84
numpy==1.26.4
pyserial==3.5
twilio==9.2.3
requests==2.31.0
werkzeug==3.0.3
itsdangerous==2.2.0
python-dotenv==1.0.0

# Use official mediapipe on most platforms (works on Python 3.10/3.11)
mediapipe==0.10.14; platform_system != "Darwin" or platform_machine != "arm64"
# Use Apple Silicon wheel providing the same `mediapipe` module
mediapipe-silicon==0.9.2.1; platform_system == "Darwin" and platform_machine == "arm64"

